% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_functions.R
\docType{package}
\name{analyticsreporting_googleAuthR}
\alias{analyticsreporting_googleAuthR}
\alias{analyticsreporting_googleAuthR-package}
\title{Google Analytics Reporting API
Accesses Analytics report data.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:34:38
filename: /Users/mark/dev/R/autoGoogleAPI/googleanalyticsreportingv4.auto/R/analyticsreporting_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/analytics.readonly
\item https://www.googleapis.com/auth/analytics
}
}

