% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{Cohort}
\alias{Cohort}
\title{Cohort Object}
\usage{
Cohort(type = NULL, dateRange = NULL, name = NULL)
}
\arguments{
\item{type}{Type of the cohort}

\item{dateRange}{This is used for `FIRST_VISIT_DATE` cohort, the cohort selects users}

\item{name}{A unique name for the cohort}
}
\value{
Cohort object
}
\description{
Cohort Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Defines a cohort. A cohort is a group of users who share a commoncharacteristic. For example, all users with the same acquisition datebelong to the same cohort.
}

