% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{ColumnHeader}
\alias{ColumnHeader}
\title{ColumnHeader Object}
\usage{
ColumnHeader(dimensions = NULL, metricHeader = NULL)
}
\arguments{
\item{dimensions}{The dimension names in the response}

\item{metricHeader}{Metric headers for the metrics in the response}
}
\value{
ColumnHeader object
}
\description{
ColumnHeader Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Column headers.
}
