% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{DateRange}
\alias{DateRange}
\title{DateRange Object}
\usage{
DateRange(startDate = NULL, endDate = NULL)
}
\arguments{
\item{startDate}{The start date for the query in the format `YYYY-MM-DD`}

\item{endDate}{The end date for the query in the format `YYYY-MM-DD`}
}
\value{
DateRange object
}
\description{
DateRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A contiguous set of days: startDate, startDate + 1 day, ..., endDate.The start and end dates are specified in[ISO8601](https://en.wikipedia.org/wiki/ISO_8601) date format `YYYY-MM-DD`.
}

