% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{DateRangeValues}
\alias{DateRangeValues}
\title{DateRangeValues Object}
\usage{
DateRangeValues(values = NULL, pivotValueRegions = NULL)
}
\arguments{
\item{values}{Each value corresponds to each Metric in the request}

\item{pivotValueRegions}{The values of each pivot region}
}
\value{
DateRangeValues object
}
\description{
DateRangeValues Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Used to return a list of metrics for a single DateRange / dimensioncombination
}
