% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{Dimension}
\alias{Dimension}
\title{Dimension Object}
\usage{
Dimension(histogramBuckets = NULL, name = NULL)
}
\arguments{
\item{histogramBuckets}{If non-empty, we place dimension values into buckets after string to}

\item{name}{Name of the dimension to fetch, for example `ga:browser`}
}
\value{
Dimension object
}
\description{
Dimension Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Dimensions](https://support.google.com/analytics/answer/1033861)are attributes of your data. For example, the dimension `ga:city`indicates the city, for example, 'Paris' or 'New York', from whicha session originates.
}

