% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{DimensionFilter}
\alias{DimensionFilter}
\title{DimensionFilter Object}
\usage{
DimensionFilter(dimensionName = NULL, operator = NULL,
  caseSensitive = NULL, expressions = NULL, not = NULL)
}
\arguments{
\item{dimensionName}{The dimension to filter on}

\item{operator}{How to match the dimension to the expression}

\item{caseSensitive}{Should the match be case sensitive? Default is false}

\item{expressions}{Strings or regular expression to match against}

\item{not}{Logical `NOT` operator}
}
\value{
DimensionFilter object
}
\description{
DimensionFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dimension filter specifies the filtering options on a dimension.
}

