% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{DimensionFilterClause}
\alias{DimensionFilterClause}
\title{DimensionFilterClause Object}
\usage{
DimensionFilterClause(operator = NULL, filters = NULL)
}
\arguments{
\item{operator}{The operator for combining multiple dimension filters}

\item{filters}{The repeated set of filters}
}
\value{
DimensionFilterClause object
}
\description{
DimensionFilterClause Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A group of dimension filters. Set the operator value to specify howthe filters are logically combined.
}

