% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{DynamicSegment}
\alias{DynamicSegment}
\title{DynamicSegment Object}
\usage{
DynamicSegment(sessionSegment = NULL, name = NULL, userSegment = NULL)
}
\arguments{
\item{sessionSegment}{Session Segment to select sessions to include in the segment}

\item{name}{The name of the dynamic segment}

\item{userSegment}{User Segment to select users to include in the segment}
}
\value{
DynamicSegment object
}
\description{
DynamicSegment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dynamic segment definition for defining the segment within the request.A segment can select users, sessions or both.
}

