% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{GetReportsRequest}
\alias{GetReportsRequest}
\title{GetReportsRequest Object}
\usage{
GetReportsRequest(reportRequests = NULL)
}
\arguments{
\item{reportRequests}{Requests, each request will have a separate response}
}
\value{
GetReportsRequest object
}
\description{
GetReportsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The batch request containing multiple report request.
}
\seealso{
Other GetReportsRequest functions: \code{\link{reports.batchGet}}
}

