% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{GetReportsResponse}
\alias{GetReportsResponse}
\title{GetReportsResponse Object}
\usage{
GetReportsResponse(reports = NULL)
}
\arguments{
\item{reports}{Responses corresponding to each of the request}
}
\value{
GetReportsResponse object
}
\description{
GetReportsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The main response class which holds the reports from the Reporting API`batchGet` call.
}
