% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{Metric}
\alias{Metric}
\title{Metric Object}
\usage{
Metric(alias = NULL, formattingType = NULL, expression = NULL)
}
\arguments{
\item{alias}{An alias for the metric expression is an alternate name for the}

\item{formattingType}{Specifies how the metric expression should be formatted, for example}

\item{expression}{A metric expression in the request}
}
\value{
Metric object
}
\description{
Metric Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Metrics](https://support.google.com/analytics/answer/1033861)are the quantitative measurements. For example, the metric `ga:users`indicates the total number of users for the requested time period.
}

