% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{MetricFilterClause}
\alias{MetricFilterClause}
\title{MetricFilterClause Object}
\usage{
MetricFilterClause(operator = NULL, filters = NULL)
}
\arguments{
\item{operator}{The operator for combining multiple metric filters}

\item{filters}{The repeated set of filters}
}
\value{
MetricFilterClause object
}
\description{
MetricFilterClause Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a group of metric filters.Set the operator value to specify how the filters are logically combined.
}

