% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{MetricHeader}
\alias{MetricHeader}
\title{MetricHeader Object}
\usage{
MetricHeader(metricHeaderEntries = NULL, pivotHeaders = NULL)
}
\arguments{
\item{metricHeaderEntries}{Headers for the metrics in the response}

\item{pivotHeaders}{Headers for the pivots in the response}
}
\value{
MetricHeader object
}
\description{
MetricHeader Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The headers for the metrics.
}
