% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{MetricHeaderEntry}
\alias{MetricHeaderEntry}
\title{MetricHeaderEntry Object}
\usage{
MetricHeaderEntry(name = NULL, type = NULL)
}
\arguments{
\item{name}{The name of the header}

\item{type}{The type of the metric, for example `INTEGER`}
}
\value{
MetricHeaderEntry object
}
\description{
MetricHeaderEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Header for the metrics.
}
