% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{OrFiltersForSegment}
\alias{OrFiltersForSegment}
\title{OrFiltersForSegment Object}
\usage{
OrFiltersForSegment(segmentFilterClauses = NULL)
}
\arguments{
\item{segmentFilterClauses}{List of segment filters to be combined with a `OR` operator}
}
\value{
OrFiltersForSegment object
}
\description{
OrFiltersForSegment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of segment filters in the `OR` group are combined with the logical ORoperator.
}
