% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{PivotHeader}
\alias{PivotHeader}
\title{PivotHeader Object}
\usage{
PivotHeader(pivotHeaderEntries = NULL, totalPivotGroupsCount = NULL)
}
\arguments{
\item{pivotHeaderEntries}{A single pivot section header}

\item{totalPivotGroupsCount}{The total number of groups for this pivot}
}
\value{
PivotHeader object
}
\description{
PivotHeader Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The headers for each of the pivot sections defined in the request.
}
