% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{PivotValueRegion}
\alias{PivotValueRegion}
\title{PivotValueRegion Object}
\usage{
PivotValueRegion(values = NULL)
}
\arguments{
\item{values}{The values of the metrics in each of the pivot regions}
}
\value{
PivotValueRegion object
}
\description{
PivotValueRegion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The metric values in the pivot region.
}
