% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{ReportRow}
\alias{ReportRow}
\title{ReportRow Object}
\usage{
ReportRow(dimensions = NULL, metrics = NULL)
}
\arguments{
\item{dimensions}{List of requested dimensions}

\item{metrics}{List of metrics for each requested DateRange}
}
\value{
ReportRow object
}
\description{
ReportRow Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A row in the report.
}

