% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_functions.R
\name{reports.batchGet}
\alias{reports.batchGet}
\title{Returns the Analytics data.}
\usage{
reports.batchGet(GetReportsRequest)
}
\arguments{
\item{GetReportsRequest}{The \link{GetReportsRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/analytics
\item https://www.googleapis.com/auth/analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/analytics, https://www.googleapis.com/auth/analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/analytics/devguides/reporting/core/v4/}{Google Documentation}

Other GetReportsRequest functions: \code{\link{GetReportsRequest}}
}

