% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{Segment}
\alias{Segment}
\title{Segment Object}
\usage{
Segment(dynamicSegment = NULL, segmentId = NULL)
}
\arguments{
\item{dynamicSegment}{A dynamic segment definition in the request}

\item{segmentId}{The segment ID of a built-in or custom segment, for example `gaid::-3`}
}
\value{
Segment object
}
\description{
Segment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The segment definition, if the report needs to be segmented.A Segment is a subset of the Analytics data. For example, of the entireset of users, one Segment might be users from a particular country or city.
}
