% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{SegmentDefinition}
\alias{SegmentDefinition}
\title{SegmentDefinition Object}
\usage{
SegmentDefinition(segmentFilters = NULL)
}
\arguments{
\item{segmentFilters}{A segment is defined by a set of segment filters which are combined}
}
\value{
SegmentDefinition object
}
\description{
SegmentDefinition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
SegmentDefinition defines the segment to be a set of SegmentFilters whichare combined together with a logical `AND` operation.
}
