% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{SegmentDimensionFilter}
\alias{SegmentDimensionFilter}
\title{SegmentDimensionFilter Object}
\usage{
SegmentDimensionFilter(caseSensitive = NULL, minComparisonValue = NULL,
  maxComparisonValue = NULL, dimensionName = NULL, operator = NULL,
  expressions = NULL)
}
\arguments{
\item{caseSensitive}{Should the match be case sensitive, ignored for `IN_LIST` operator}

\item{minComparisonValue}{Minimum comparison values for `BETWEEN` match type}

\item{maxComparisonValue}{Maximum comparison values for `BETWEEN` match type}

\item{dimensionName}{Name of the dimension for which the filter is being applied}

\item{operator}{The operator to use to match the dimension with the expressions}

\item{expressions}{The list of expressions, only the first element is used for all operators}
}
\value{
SegmentDimensionFilter object
}
\description{
SegmentDimensionFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dimension filter specifies the filtering options on a dimension.
}
