% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{SegmentFilter}
\alias{SegmentFilter}
\title{SegmentFilter Object}
\usage{
SegmentFilter(sequenceSegment = NULL, not = NULL, simpleSegment = NULL)
}
\arguments{
\item{sequenceSegment}{Sequence conditions consist of one or more steps, where each step is}

\item{not}{If true, match the complement of simple or sequence segment}

\item{simpleSegment}{A Simple segment conditions consist of one or more dimension/metric}
}
\value{
SegmentFilter object
}
\description{
SegmentFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
SegmentFilter defines the segment to be either a simple or a sequencesegment. A simple segment condition contains dimension and metric conditionsto select the sessions or users. A sequence segment condition can be used toselect users or sessions based on sequential conditions.
}
