% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{SegmentFilterClause}
\alias{SegmentFilterClause}
\title{SegmentFilterClause Object}
\usage{
SegmentFilterClause(dimensionFilter = NULL, metricFilter = NULL,
  not = NULL)
}
\arguments{
\item{dimensionFilter}{Dimension Filter for the segment definition}

\item{metricFilter}{Metric Filter for the segment definition}

\item{not}{Matches the complement (`!`) of the filter}
}
\value{
SegmentFilterClause object
}
\description{
SegmentFilterClause Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Filter Clause to be used in a segment definition, can be wither a metric ora dimension filter.
}

