% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{SegmentMetricFilter}
\alias{SegmentMetricFilter}
\title{SegmentMetricFilter Object}
\usage{
SegmentMetricFilter(metricName = NULL, scope = NULL,
  maxComparisonValue = NULL, comparisonValue = NULL, operator = NULL)
}
\arguments{
\item{metricName}{The metric that will be filtered on}

\item{scope}{Scope for a metric defines the level at which that metric is defined}

\item{maxComparisonValue}{Max comparison value is only used for `BETWEEN` operator}

\item{comparisonValue}{The value to compare against}

\item{operator}{Specifies is the operation to perform to compare the metric}
}
\value{
SegmentMetricFilter object
}
\description{
SegmentMetricFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metric filter to be used in a segment filter clause.
}
