% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{SegmentSequenceStep}
\alias{SegmentSequenceStep}
\title{SegmentSequenceStep Object}
\usage{
SegmentSequenceStep(matchType = NULL, orFiltersForSegment = NULL)
}
\arguments{
\item{matchType}{Specifies if the step immediately precedes or can be any time before the}

\item{orFiltersForSegment}{A sequence is specified with a list of Or grouped filters which are}
}
\value{
SegmentSequenceStep object
}
\description{
SegmentSequenceStep Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A segment sequence definition.
}

