% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{SequenceSegment}
\alias{SequenceSegment}
\title{SequenceSegment Object}
\usage{
SequenceSegment(firstStepShouldMatchFirstHit = NULL,
  segmentSequenceSteps = NULL)
}
\arguments{
\item{firstStepShouldMatchFirstHit}{If set, first step condition must match the first hit of the visitor (in}

\item{segmentSequenceSteps}{The list of steps in the sequence}
}
\value{
SequenceSegment object
}
\description{
SequenceSegment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Sequence conditions consist of one or more steps, where each step is definedby one or more dimension/metric conditions. Multiple steps can be combinedwith special sequence operators.
}

