% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{SimpleSegment}
\alias{SimpleSegment}
\title{SimpleSegment Object}
\usage{
SimpleSegment(orFiltersForSegment = NULL)
}
\arguments{
\item{orFiltersForSegment}{A list of segment filters groups which are combined with logical `AND`}
}
\value{
SimpleSegment object
}
\description{
SimpleSegment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Simple segment conditions consist of one or more dimension/metricconditions that can be combined.
}

