% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{DimensionFilter}
\alias{DimensionFilter}
\title{DimensionFilter Object}
\usage{
DimensionFilter(caseSensitive = NULL, dimensionName = NULL,
  operator = NULL, not = NULL, expressions = NULL)
}
\arguments{
\item{caseSensitive}{Should the match be case sensitive? Default is false}

\item{dimensionName}{The dimension to filter on}

\item{operator}{How to match the dimension to the expression}

\item{not}{Logical `NOT` operator}

\item{expressions}{Strings or regular expression to match against}
}
\value{
DimensionFilter object
}
\description{
DimensionFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dimension filter specifies the filtering options on a dimension.
}
