% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{DimensionFilterClause}
\alias{DimensionFilterClause}
\title{Google Analytics Reporting API Objects 
Accesses Analytics report data.}
\usage{
DimensionFilterClause(operator = NULL, filters = NULL)
}
\arguments{
\item{operator}{The operator for combining multiple dimension filters}

\item{filters}{The repeated set of filters}
}
\value{
DimensionFilterClause object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:25:36
filename: /Users/mark/dev/R/autoGoogleAPI/googleanalyticsreportingv4.auto/R/analyticsreporting_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
DimensionFilterClause Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A group of dimension filters. Set the operator value to specify howthe filters are logically combined.
}
