% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{OrderBy}
\alias{OrderBy}
\title{OrderBy Object}
\usage{
OrderBy(sortOrder = NULL, fieldName = NULL, orderType = NULL)
}
\arguments{
\item{sortOrder}{The sorting order for the field}

\item{fieldName}{The field which to sort by}

\item{orderType}{The order type}
}
\value{
OrderBy object
}
\description{
OrderBy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies the sorting options.
}
