% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{PivotHeader}
\alias{PivotHeader}
\title{Google Analytics Reporting API Objects 
Accesses Analytics report data.}
\usage{
PivotHeader(totalPivotGroupsCount = NULL, pivotHeaderEntries = NULL)
}
\arguments{
\item{totalPivotGroupsCount}{The total number of groups for this pivot}

\item{pivotHeaderEntries}{A single pivot section header}
}
\value{
PivotHeader object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 22:34:38
filename: /Users/mark/dev/R/autoGoogleAPI/googleanalyticsreportingv4.auto/R/analyticsreporting_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
PivotHeader Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The headers for each of the pivot sections defined in the request.
}

