% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{PivotHeaderEntry}
\alias{PivotHeaderEntry}
\title{PivotHeaderEntry Object}
\usage{
PivotHeaderEntry(dimensionValues = NULL, dimensionNames = NULL,
  metric = NULL)
}
\arguments{
\item{dimensionValues}{The values for the dimensions in the pivot}

\item{dimensionNames}{The name of the dimensions in the pivot response}

\item{metric}{The metric header for the metric in the pivot}
}
\value{
PivotHeaderEntry object
}
\description{
PivotHeaderEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The headers for the each of the metric column corresponding to the metricsrequested in the pivots section of the response.
}
