% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{Report}
\alias{Report}
\title{Report Object}
\usage{
Report(data = NULL, nextPageToken = NULL, columnHeader = NULL)
}
\arguments{
\item{data}{Response data}

\item{nextPageToken}{Page token to retrieve the next page of results in the list}

\item{columnHeader}{The column headers}
}
\value{
Report object
}
\description{
Report Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The data response corresponding to the request.
}
