% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{ReportData}
\alias{ReportData}
\title{ReportData Object}
\usage{
ReportData(rowCount = NULL, samplingSpaceSizes = NULL, maximums = NULL,
  samplesReadCounts = NULL, minimums = NULL, rows = NULL, totals = NULL,
  isDataGolden = NULL)
}
\arguments{
\item{rowCount}{Total number of matching rows for this query}

\item{samplingSpaceSizes}{If the results are}

\item{maximums}{Minimum and maximum values seen over all matching rows}

\item{samplesReadCounts}{If the results are}

\item{minimums}{Minimum and maximum values seen over all matching rows}

\item{rows}{There's one ReportRow for every unique combination of dimensions}

\item{totals}{For each requested date range, for the set of all rows that match}

\item{isDataGolden}{Indicates if response to this request is golden or not}
}
\value{
ReportData object
}
\description{
ReportData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The data part of the report.
}

