% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticsreporting_objects.R
\name{ReportRow}
\alias{ReportRow}
\title{ReportRow Object}
\usage{
ReportRow(metrics = NULL, dimensions = NULL)
}
\arguments{
\item{metrics}{List of metrics for each requested DateRange}

\item{dimensions}{List of requested dimensions}
}
\value{
ReportRow object
}
\description{
ReportRow Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A row in the report.
}
